feat: renamed files and minor fixes
