+++ ui/src/modals/RuleModalInputTabCard.vue
