+++ ui/src/modals/RuleModalOutputTabCardBody.vue
