+++ ui/src/modals/ezbidsimporter.vue
@@ -2,7 +2,7 @@
 <b-modal :no-close-on-backdrop="true" title="Import ezBIDS Dataset" ref="modal" size="lg" @ok="submit">
     <div v-if="!task">
         <b-row>
-           <b-col class="text-muted" cols="3">ezBIDS Session ID</b-col>
+            <b-col class="text-muted" cols="3">ezBIDS Session ID</b-col>
             <b-col>
                 <b-form-input v-model="sessionId" readonly></b-form-input>
                 <br>
@@ -112,11 +112,12 @@ export default {
             this.task = null;
             
             //load finalized ezbids content
-            this.$http.get(Vue.config.ezbids_api+"/download/"+this.sessionId+"/finalized.json").then(res=>{
+            this.$http.get(Vue.config.ezbids_api+"/download/"+this.sessionId+"/finalized.json").then((res) => {
                 this.ezBIDS = res.data;
                 this.project_name = this.ezBIDS.datasetDescription.Name;
                 this.project_desc = this.ezBIDS.readme;
-            }).catch(err=>{
+            }).catch((err) => {
+                console.error(err);
                 this.$notify({type: 'error', text: "Failed to load the specified ezBIDS session. Please contact brainlife.io team"});
                 this.close();
             });
@@ -129,12 +130,12 @@ export default {
                         service: "brainlife/app-bids-import",
                     })
                 },
-            }).then(res=>{
+            }).then((res) => {
                 if(res.data.tasks.length != 0) {
                     this.task = res.data.tasks[res.data.tasks.length-1];
                     this.subscribeTask(this.task);
                 }
-            }).catch(err=>{
+            }).catch((err) => {
                 console.error(err);
                 this.$notify({type: 'error', text: err.body.message});
             });
@@ -159,10 +160,14 @@ export default {
             this.$http.post('project', {
                 name: this.project_name, 
                 desc: this.project_desc,
-            }).then(res=>{
+            }).then((res) => {
                 this.project = res.data._id;
+                if (this.pipelineName) {
+                    console.log('EMIT', this.project, this.pipelineName)
+                    this.$root.$emit("create_pipeline", { projectId: this.project, pipelineName: this.pipelineName })
+                }
                 this.$root.$emit("refresh_jwt", this.submit_import);
-            }).catch(err=>{
+            }).catch((err) => {
                 console.error(err);
                 this.$notify({type: "error", text: err.response.data.message});
             });
@@ -183,14 +188,10 @@ export default {
                         jwt: Vue.config.jwt,
                     }
                 }
-                this.$http.post(Vue.config.amaretti_api+'/task', params).then(res=>{
+                this.$http.post(Vue.config.amaretti_api+'/task', params).then((res) => {
                     this.task = res.data.task;
                     this.subscribeTask(this.task);
-                    if (this.pipelineName) {
-                        this.$root.$emit("create_pipeline", { projectId: this.project, pipelineName: this.pipelineName })
-                    }
-                    this.close();
-                }).catch(err=>{
+                }).catch((err) => {
                     console.error(err);
                     this.$notify({type: 'error', text: err.body.message});
                 });
@@ -209,10 +210,10 @@ export default {
                 this.$http.post(Vue.config.amaretti_api+'/instance', {
                     name,
                     group_id: this.project.group_id,
-                }).then(res=>{
+                }).then((res) => {
                     const instance = res.data;
                     return cb(null, instance);
-                }).catch(err=>{
+                }).catch((err) => {
                     console.error(err);
                 });
             });
@@ -244,11 +245,11 @@ export default {
 
         cancelImport() {
             this.$http.put(Vue.config.amaretti_api+'/task/stop/'+this.task._id)
-            .then(res=>{
+            .then((res) => {
                 this.$notify({ text: res.data.message, type: 'success'});
                 this.close();
             })
-            .catch(err=>{
+            .catch((err) => {
                 console.error(err); 
             });
         },
